+++ paddle_billing/Entities/Subscriptions/SubscriptionAdjustmentItem.py
@@ -9,7 +9,7 @@ class SubscriptionAdjustmentItem:
     item_id: str
     type: AdjustmentType
     amount: str | None
-    proration: Proration
+    proration: Proration | None
     totals: AdjustmentItemTotals
 
     @staticmethod
@@ -18,6 +18,6 @@ class SubscriptionAdjustmentItem:
             item_id=data["item_id"],
             type=AdjustmentType(data["type"]),
             amount=data.get("amount"),
-            proration=Proration.from_dict(data["proration"]),
+            proration=Proration.from_dict(data["proration"]) if data.get("proration") else None,
             totals=AdjustmentItemTotals.from_dict(data["totals"]),
         )