+++ paddle_billing/Resources/Transactions/Operations/Create/TransactionCreateItemWithPrice.py
@@ -1,8 +1,8 @@
 from __future__ import annotations
 from dataclasses import dataclass
 
-from paddle_billing.Entities.Transactions.TransactionNonCatalogPrice import TransactionNonCatalogPrice
-from paddle_billing.Entities.Transactions.TransactionNonCatalogPriceWithProduct import (
+from paddle_billing.Resources.Transactions.Operations.Price.TransactionNonCatalogPrice import TransactionNonCatalogPrice
+from paddle_billing.Resources.Transactions.Operations.Price.TransactionNonCatalogPriceWithProduct import (
     TransactionNonCatalogPriceWithProduct,
 )