+++ paddle_billing/Resources/Transactions/Operations/Update/TransactionUpdateItemWithPrice.py
@@ -1,0 +1,13 @@
+from __future__ import annotations
+from dataclasses import dataclass
+
+from paddle_billing.Resources.Transactions.Operations.Price.TransactionNonCatalogPrice import TransactionNonCatalogPrice
+from paddle_billing.Resources.Transactions.Operations.Price.TransactionNonCatalogPriceWithProduct import (
+    TransactionNonCatalogPriceWithProduct,
+)
+
+
+@dataclass
+class TransactionUpdateItemWithPrice:
+    price: TransactionNonCatalogPrice | TransactionNonCatalogPriceWithProduct
+    quantity: int